Fix issue #55.  Crawl now replaces activities.
Previously there would be lots of db constraint errors during a crawl
when we attempted to remove the activity objects related to a
resource. These have been removed by correcting the on delete cascade
behaviour of some foriegn keys.

Parse errors are now logged into the `Log` table in the database
which can be cleared out withe the new `cleanup` command.