more relaxed parser #17
allow policy-code to be nullable. Relax the parser for policy-marker,
hierarchy, sector-percentages, recipient-region, recipient-country.
Exceptions now handled/logged lower down in the parser so that an
activity will not fail to parse if any of the above fail to parse, or in
many-to-one_activity, only fail the element that failed, not the entire
list.

Bad codes in codelists now handled in from_codelist instead of
being thrown from from_codelists

Logging code needs to be refactored to be cleaner.