[#237] Prevent 500 status on policy-marker filter
This prevents the server from returning a 500 status when using
the  api filter.

The policy-marker filter causes strings to be passed around
internally, rather than converting to an enum as with other
filters. As such, this is a bandaid patch and does not fix or
answer the underlying query of why an enum is not reaching
this point in the code.

It may also be the case that other types, such as an int could
reach this point in the code, however this has not been fully
tested and/or investigated.