Update psycopg2 version
Updates psycopg2 to the latest version. This is needed as Travis runs a later version of postgres but the with an incompatible version. This causes all travis builds to fail, for example: https://travis-ci.org/IATI/IATI-Datastore/builds/327278368  It appears that requires.io is not properly working, which awaits further investigation.