Update base URLs for the Registry
As part of the migration of the IATI Registry to a new supplier, the Registry is now hosted at https://iatiregistry.org with 301 redirects from URLs prepended with `http` and/or `www`.

However, the ckanapi package cannot handle 301 redirects, which caused ValidationError exceptions on the `301 Moved Permanently` content that was returned. This commit updates the URLs so that expected content is returned.